Add functionality for page refresh
A function was added to disable all inputs and
mouse events, if the form submit was successful,
until after the page refreshes with the 5 second
delay. There was also some coded added to the
submit event listener that deals with the radios.
It clears the background color. Before the
background color would stay in the radio container
but everything else reset. This fixes that.