Fix padding issues and success message issues
There was an issue centering the text and filling
up the mobile screen with the success message.
This was corrected. There was some issues with
padding in the first name name input being
slightly larger than needed, giving the layout
an unbalanced look. This was corrected to match
the other inputs.